+++ @navikt/aksel-icons/config/figma/paginate.mjs
@@ -1,7 +1,10 @@
 export function paginate(arr, size) {
   return arr.reduce((acc, val, i) => {
     let idx = Math.floor(i / size);
-    let page = acc[idx] || (acc[idx] = []);
+    if (!acc[idx]) {
+      acc[idx] = [];
+    }
+    let page = acc[idx];
     page.push(val);
 
     return acc;