+++ @navikt/core/react/src/layout/box/Box.stories.tsx
@@ -1,11 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import React, { ReactNode } from "react";
+import React from "react";
 import { ChevronRightIcon } from "@navikt/aksel-icons";
 import { BodyLong, Detail, Heading } from "../../typography";
 import { HGrid } from "../grid";
 import { HStack, VStack } from "../stack";
-import { BackgroundToken, BorderRadiiToken } from "../utilities/types";
-import { Box } from "./Box";
+import { Box, BoxProps } from "./Box";
 
 const meta: Meta<typeof Box> = {
   title: "ds-react/Primitives/Box",
@@ -34,11 +33,7 @@ const Card = ({
   background,
   borderRadius = "xlarge",
   children,
-}: {
-  background?: BackgroundToken;
-  borderRadius?: BorderRadiiToken;
-  children: ReactNode;
-}) => (
+}: Pick<BoxProps, "background" | "borderRadius" | "children">) => (
   <Box
     padding="4"
     background={background}
@@ -373,14 +368,36 @@ export const BorderRadius = () => (
   </VStack>
 );
 
+export const MarginDemo = () => (
+  <VStack gap="4">
+    <Box borderWidth="1">
+      <Box
+        background="surface-info"
+        margin="5"
+        marginInline={{ xs: "20", lg: "10" }}
+      >
+        Box
+      </Box>
+    </Box>
+  </VStack>
+);
+
 export const PaddingDemo = () => (
   <VStack gap="4">
     <Box
       background="bg-subtle"
-      padding="0"
+      padding="5"
       paddingInline={{ xs: "20", lg: "10" }}
     >
       Box
+    </Box>
+  </VStack>
+);
+
+export const AsChild = () => (
+  <VStack gap="4">
+    <Box borderRadius="large" padding="4" asChild>
+      <button onClick={() => alert("clicked")}>Box is now a button</button>
     </Box>
   </VStack>
 );