+++ aksel.nav.no/website/components/website-modules/preview/PagePreview.tsx
@@ -1,0 +1,8 @@
+import dynamic from "next/dynamic";
+import { PreviewProps } from "./Preview.types";
+
+const LiveQueryProvider = dynamic(() => import("./LiveQueryProvider"));
+
+export function PagePreview(props: PreviewProps) {
+  return <LiveQueryProvider {...props} />;
+}