+++ aksel.nav.no/website/pages/prinsipper/[...prinsipp].tsx
@@ -1,6 +1,5 @@
 import cl from "clsx";
 import { GetServerSideProps } from "next/types";
-import { Suspense, lazy } from "react";
 import { BodyLong, BodyShort, Heading, Label } from "@navikt/ds-react";
 import Bilde from "@/cms/bilde/Bilde";
 import Footer from "@/layout/footer/Footer";
@@ -16,6 +15,7 @@ import {
   TableOfContentsT,
 } from "@/types";
 import { abbrName, dateStr, generateTableOfContents } from "@/utils";
+import { PagePreview } from "@/web/preview/PagePreview";
 import { SEO } from "@/web/seo/SEO";
 import TableOfContents from "@/web/toc/TableOfContents";
 import NotFotfund from "../404";
@@ -31,13 +31,10 @@ export const query = `{
   {
     ...,
     "slug": slug.current,
-    "content": select(
-      $valid == "true" => content[]{
-        ...,
-        ${destructureBlocks}
-      },
-      $valid != "true" => []
-    ),
+    content[]{
+      ...,
+      ${destructureBlocks}
+    },
     ${contributorsAll}
   }
 }`;
@@ -50,7 +47,6 @@ export const getServerSideProps: GetServerSideProps = async (
 
   const { prinsipp } = await getClient().fetch(query, {
     slug: `prinsipper/${slugString}`,
-    valid: "true",
   });
 
   return {
@@ -200,26 +196,31 @@ const Page = ({ prinsipp: data, publishDate, toc }: PageProps["props"]) => {
   );
 };
 
-const WithPreview = lazy(() => import("@/preview"));
-
-const Wrapper = (props: any) => {
-  if (props?.preview) {
-    return (
-      <Suspense fallback={<Page {...props} />}>
-        <WithPreview
-          comp={Page}
-          query={query}
-          props={props}
-          params={{
-            slug: `prinsipper/${props.slug}`,
-            valid: "true",
-          }}
-        />
-      </Suspense>
-    );
-  }
-
-  return <Page {...props} />;
-};
-
-export default Wrapper;
+export default function PrinsippPage(props: PageProps["props"]) {
+  return props.preview ? (
+    <PagePreview
+      query={query}
+      props={props}
+      params={{
+        slug: `prinsipper/${props.slug}`,
+      }}
+    >
+      {(previewProps, loading) => {
+        if (loading) {
+          return <Page {...props} />;
+        }
+        return (
+          <Page
+            {...previewProps}
+            toc={generateTableOfContents({
+              content: previewProps?.prinsipp?.content,
+              type: "aksel_prinsipp",
+            })}
+          />
+        );
+      }}
+    </PagePreview>
+  ) : (
+    <Page {...props} />
+  );
+}