+++ aksel.nav.no/website/sanity/interface/client.server.ts
@@ -1,5 +1,5 @@
 import { createClient } from "next-sanity";
-import { clientConfig } from "../../sanity/config";
+import { clientConfig } from "../config";
 
 // Set up the client for fetching data in the getProps page functions
 export const sanityClient = createClient({
@@ -12,10 +12,18 @@ export const sanityClient = createClient({
 // Set up a preview client with serverless authentication for drafts
 /* Should not be needed anymore after migration of preview solution */
 
-export const noCdnClient = (token) =>
+export const noCdnClient = (token: string) =>
   createClient({
     ...clientConfig,
     token,
+  });
+
+export const previewDraftsClient = () =>
+  sanityClient.withConfig({
+    useCdn: false,
+    ignoreBrowserTokenWarning: true,
+    perspective: "previewDrafts",
+    withCredentials: true,
   });
 
 // Helper function for easily switching between normal client and preview client